Complete server side handling of report submission
The handling of the /report route should store the provided telemetry
in the telemetry DB. Leverages the telemetry extractor to process the
incoming report and extract the bundles and corresponding data items.

Integrated server config functionality with the application.

Moved database and table definitions to separate files.

Updated README.md to reflect location change of the git repositories.

TODO: We should only extract bundles from reports, and should handle
the extraction of data items from bundles as part of adding data items
to the telemetry DB; this is required because we need info from the
bundle header when adding data items to the DB.