Improve telemetry data storage handling workflow.
Remove the staging of data items in the data store and instead add
data items to the Telemetry DB as part of processing bundles. This
is required because we need info from the bundle header when adding
data items.

Cleanup the telemetryData, tagElements and tagLists tables related
definitions and helpers.

Fixed some errors in the handling of tags.