Add support for CI testing of PRs
Add a top-level Makefile that can drive testing and related actions.

This top-level Makefile itself implements a simple wrapper for the
configured target actions that invokes them via a recursive make in
the specified subdirectories.

A Makefile.subdir is provided that will run the target actions, and
should be symlinked into the relevant subdirectories.

Update README.md with instructions on how to run tests, and cleanup
some stale text.

A GitHub Action CI workflow is defined that will trigger the tests
to be run when a PR is proposed or updated.

Caveat: Currently uses local checkouts for testing purposes

Closes: #8