Add support for telemetry type specific tables
Using the existing SCC hardware info blob as an example, implement a
telemetry type specific transformation to store the data in the table
specific to that telemetry type.

This new mechanism define row handlers matching the associated tables
that can be used to process the received JSON blob and store it into
the appropriate telemetry table, with a default handler that stores
the JSON blob as a single field.

Reorganise the app code to move logically related code into the same
files, and renamed some files to better reflect their content.

Updated the staging and subsequent processing of received telemetry
reports to align with the expected background processing model that
will be implemented at a later date.