SQL value placeholder generators
Depending on which DB backend is being used, we need to use different
placeholder values in SQL statements. The new QuestionMarker() and
DollarCounter() placeholder generators with generate appropriate
placeholder values for use with SQLite and PostgreSQL respectively.

Also added a Makefile to the app/ directory, and added app to the
list of subdirectorys that the top-level Makefile targets.