Add telemetry-admin placeholder
The telemetry-admin service will be used to perform management
operations on the databases managed by the telemetry-server service.

For now those management operations are still being defined so this
patch set is just adding support for building and running an admin
service that responds to healthchecks.

Note also that support has been added for API requests to report
their success status at debug rather than info level, with the
healthcheck request handlers being updated to make use of this
so that the frequent healthcheck operations don't cause too much
chatter in the service logs.

The Dockerfile has been updated to support building both the server
and admin services.

The docker/compose.yaml has been updated to start the admin service
as a dependency of the main telemetry service.

Updated README.md to reflect addition of telemetry-admin service
to the docker compose deployment.

NOTE: The helm charts will be updated to support deploying the admin
service is a followup PR.