HTTP 401 responses should include WWW-Authentciate header
When generating a HTTP 401 StatusUnauthorized response to a telemetry
client the response should include an appropriate WWW-Authenticate
header indicating to the client what action it should take to address
it's lack of authorization.

The generated WWW-Authenticate header uses a custom "Bearer" challenge
specifying a realm of "suse-telemetry-service" with a scope of either
"authenticate" or "register" indicating the appropriate action which
the client needs to take.

Fixes: #47