Address Customer identification mapping (#66)
Add a new customers table which is used to manage customerIds that
are provided in telemetry submissions.

Update the telemetry data storage handling to obtain the appropriate
customerRefId for a bundle's customer id.

Update the database table definition and management routines with
support for specifying foreign keys via a new ForeignKeys field in
the TableSpec structure.

Restore the telemetrydata table foreign key definition for tagSetId
targetting the id field of the tagsets table.

Update telemetrydata table definition to replace the customerId
field with a customerRefId field that is a foreign key reference
to the associated id field in the new customers table.

Ensure failures adding a report to the reports table are handled
appropriately, resulting in a failed /report request.

Fix ClientsRow.Exists() error message to match what the routine is
doing.

Make target enhancements:
* generate now support LOG_LEVEL specification
* test-verbose shows detailed coverage statistics

Add a dependency for the Telemetry Admin service on the Telemetry
Gateway service in the docker/compose.yaml to avoid racing attempts
to initialise the table definitions in the backend postgres database.