Allow duplicate clientIds in clients table (#67)
Duplicate registrations will result in an HTTP StatusConflict (409)
but for now just log a warning in the server logs if we ever see a
duplicate clientId as part of the registration value in a /register
request.

Duplicate UUIDs being generated by independent systems should be a
rare occurrence, and the risk should be further reduced for client
systems that are registered with the SCC, as they will have their
telemetry client id assigned to them, which should be unique with
respect to all other registered systems at that point in time.