Use upstream telemetry repo
Instead of requiring that github.com/SUSE/telemetry be cloned on the
appropriate branch beside the telemetry-server repo, rely on normal
mod dependencies to provide the telemetry repo.

Update the go.mod to remove the replace directive for the telemetry
repo, and update the go.sum to match the current (v0.1.1) version of
the upstream telemetry repo.

Update the Dockerfile build process to optionally clone the telemetry
repo and use it as a build input if a branch or tag other than main
is specified via the build arguments.

Update Makefile rules to pass the appropriate build args for docker
and compose targets if a branch other than main is specified via the
TELEMETRY_REPO_BRANCH variable.

Update GitHub Actions ci.yaml workflow to clone and use the telemetry
repo to run the source tests if a branch other than main is specified.

Updated local-server make target rules to add support for running both
the server and tests using the locally cloned telemetry repo if found.
This is achieved by copying the telemetry-server repo, and, if found,
the telemetry repo cloned beside it, to a staging directory under
/tmp/telemetry/local, adjusting the copied telemetry-server repo's
go.mod settings with a appropriate as needed, and then the server or
the tests are run from within that staging directory, depending on the
specified make target.

Update README.md to reflect these changes.