Align compose postgres init with product deployment
Update the docker compose postgres DB initialisation to align with
the production deployment workflow.

Add a new init-db init container to the compose.yaml that runs after
bringing the postgres DB container up to initialise the DB similarly
to how the helm charts use an init container to init the postgres DB.

Updated the tsg service to also depend on the init-db service having
completed successfully.

Also remove the copy of the testdata into the builder base container
for the telemetry images to avoid triggering a rebuild of the layer
containing the built telemetry binaries if the testdata contents are
updated. This change also requires copying the config files from the
local dir rather than the builder image.

Updated the compose server and admin configs in the testdata area to
use the telemetry user created by the init-db service rather that the
master postgres user.