Address dependabot reported jwt dependency issues
CVE: CVE-2025-30204

GHSA: GHSA-mh63-6h87-95cp

Updated go.mod and go.sum files to pull in updated application and
testing dependencies.

Also update Makefile tooling to assist with go mod maintenance, with
mod-update and test-mod-update targets that can be used to trigger
updating to the latest standard and testing dependencies respectively.
Renamed the tidy target to mod-tidy to better align with other module
management target naming.

Additionally module management target actions will be skipped if there
is no go.mod in the current directory.

Added support for specifying GONOPROXY environment setting to the go mod
and get operations in the Makefile and Dockerfile to ensure that we use
the latest versions of relavant modules rather than the possibly stale
versions cached in the go mod proxy infrastructure.

Add a top-level telemetryserver package to allow go get to be run at
for the top-level tree.