Chart enhancements for testing (#84)
Overview of changes
===================
Update the helm charts to support deployment of both production
and testing environmets.

As part of doing so a telemetry-postgres chart has been added that
is responsible to defining the telemetry-postgres service, which
replaces the rds-postgres service previously defined by the rds-init
chart. For a production deployment the service will point to the
external PostgreSQL service that will be leveraged by our production
deployment. For a testing environment the service will point to a
local postgres instance that will be deployed by the chart. The
deployment mode is controlled by the external.enabled setting in
the chart's values.yaml file; if set to true then the external
PostgreSQL service can be specified via the external.hostname
setting in the values.yaml.

The rds-init chart is responsible for initialising the PostgreSQL
service that is pointed to by the postgres.host setting in the
chart's values.yaml file.

The telemetry-server chart works mostly the same as before, though
the config file generation has been tweaked somewhat to allow for
dynamically generating the database connection params.

Telemetry Postgres Chart
========================
Recreate the chart to deploy a local postgres instance, renaming it
to telemetry-postgres. Move the definition of the telemetry postgres
service end point into this chart and rename it from rds-postgres to
telemetry-postgres to be more generic.

The default mode of operation for the telemetry-postgres chart will
be to define the telemetry-postgres service pointing to a specified
external postgres provider. However, by overriding the defaults it
will deploy a local postgres instance, and define telemtry-postgres
as a service pointing to the local postgres instance instead.

RDS Init Chart
========================
Update the rds-init chart's scripting to use the postgres hostname
specified via the values.yaml rather than as a hardocded value.

Leverage the same scripts that the telemetry-postgres image has
builtin to seed the configMaps used to mount scripts into the
standard postgres image; future changes will switch to using the
telemetry-postgres image, removing the need for configMaps to
mount the scripts into the image.

Add a resources block to the rds-init values.yaml to satisfy the
OpenPlatform requirement for resources to be specified.

Telemetry Server Chart
========================
Update the telemetry-server chart to generate the service config
dynamically from specified telemetry settings in the values.yaml
rather than as an embedded YAML value in the values.yaml. This
both simplifies overriding config settings and allows the config
to be generated using values specified in other values.yaml blocks
such as the telemetry postgres backend setting, rather than having
to use hardcoded values. Also looks up the telemetry user password
in the rds-postgres-config.

Similarly, simplify the definition of, and thus the overriding of,
the service probe definitions in the values.yaml, using values that
are specified in the telemetry settings block.

Use the service port settings from the values.yaml to define the
port setting used for the telemetry-server service definition rather
than hardcoding them.

Update the telemetry-server chart to allow specifying the cert host
and domain via the values.yaml, and add a new helper template that
can be used to generate the fully qualified domain name for which
the telemetry-server's certificate will be defined.

Similarly to the rds-init chart, leverage the same scripts that the
telemetry-postgres image has builtin to seed the configMaps used to
mount the pre-deploy checks scripts into the initContainer's
standard postgres image; future changes will switch to using the
telemetry-postgres image, removing the need for configMaps to
mount the scripts into the image.

Add a resources block to the initContainer definition to satisfy the
OpenPlatform requirement for resources to be specified.

Misc Changes
============
Switch to using registry.suse.com/suse/postgres as our base image
for postgres as it is better maintained with critical CVE issues
being fixed promptly. Note that this changed the mount point path
for the database backing volumes to /var/lib/pgsql/data.

Add a pre-deploy-checks job into the docker compose configuration
that will exercise the pre_deploy_checks.bash script similarly to
a helm deployment of the telemetry-server chart.

Minor tweaks to the targets triggered by a 'make clean' to better
clean up stale or unused resources.

Switch GHA PR testing CI workflow to use the compose-e2e target
rather than the docker-e2e target so that we exercise all of our
deployment images.

Fix trailing whitespace issues throughout the code base.