Ensure we build and publish correct target images (#95)
When building images using our Dockerfile we need to ensure that we
specify the build stage target for the image that we want to build
using the target attribute.

Additionally enable GitHub Actions caching for the image build action
to speed up subsequent build runs.

Also enable pulling of the latest base images when building images,
either locally (via docker-build or compose-build make targets) and
in the GitHub Actions workflow.