Implement simplified /report request handling
Switch the /report request handling to process the provided telemetry
report inline, rather than staging it and then running the background
processing loop. This should reduce the operational DB's load for now
as we don't need to create a transaction to add the report.

Retain the original code flow for now in a disabled by default state,
with tweaks to return the id of the entry added to the staging table
that will be used as the processing id in the payload of the /report
response; a value of 0 indicates inline processing, and a non-zero
value indicates that background processing was used.

Added a missing return in the error handling of the /register request
in the case where there is a failure generating a new AuthToken.