DB Enhancements (#105)
Add support for columns to be specified as unique, resulting in the
creation of appropriate indices.

Add support for specifying orderby, optionally with descending, when
constructing select statements.

Add a SetTableSpec() method to TableRowCommon to setup the TableSpec
for a table row struct. Similarly add a GetTableSpec() method to get
the associated TableSpec, which will panic with an appropriate error
if the structure tablespec hasn't been initialised yet.