Ensure re-registration stores correct client data
Fix a bug introduced when doing a duplicate client id check which had
an undesirable side effect of overwriting client request provided data
fields with the values associated with the duplicate entry.

Move the duplicate client checking to a helper routine that uses it's
own ClientsRow struct, rather than re-using the one that was used to
check for the complete registration already existing.

Update the TestRegisterClient test to check for this scenario.

Fixes: TEL-251